fix:커뮤니티 리팩토링 구조 수정 (훅 분리X)
